Add spark-cypher-tck and refactor okapi-tck
