--- okapi-tck/src/test/scala/org/opencypher/okapi/tck/Tags.scala
@@ -1,30 +1,0 @@
-/*
- * Copyright (c) 2016-2018 "Neo4j, Inc." [https://neo4j.com]
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.opencypher.okapi.tck
-
-import org.scalatest.Tag
-
-object Tags {
-
-  object TckCapsTag extends Tag("TckCaps")
-
-  object TckCoscTag extends Tag("TckCosc")
-
-  object WhiteList extends Tag("WhiteList Scenario")
-
-  object BlackList extends Tag("BlackList Scenario")
-
-}
