+++ spark-cypher-tck/pom.xml
@@ -5,30 +5,23 @@
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
-    <groupId>org.opencypher</groupId>
     <artifactId>okapi</artifactId>
+    <groupId>org.opencypher</groupId>
     <version>1.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>okapi-tck</artifactId>
+  <artifactId>spark-cypher-tck</artifactId>
 
-  <name>Okapi TCK - TCK integration for the Okapi pipeline</name>
+  <name>Cypher for Apache Spark - TCK</name>
 
   <properties>
     <project.rootdir>${project.parent.basedir}</project.rootdir>
     <dep.tck.version>1.0.0-M08</dep.tck.version>
-    <dep.apiguardian-api.version>1.0.0</dep.apiguardian-api.version>
   </properties>
 
   <dependencies>
-    <!-- Test -->
 
-    <dependency>
-      <groupId>org.opencypher</groupId>
-      <artifactId>tck</artifactId>
-      <version>${dep.tck.version}</version>
-      <scope>test</scope>
-    </dependency>
+    <!-- Test -->
 
     <dependency>
       <groupId>org.opencypher</groupId>
@@ -42,17 +35,39 @@
       <groupId>org.opencypher</groupId>
       <artifactId>okapi-ir</artifactId>
       <version>${project.parent.version}</version>
+      <type>test-jar</type>
       <scope>test</scope>
     </dependency>
 
-    <!-- Workaround for https://youtrack.jetbrains.com/issue/SCL-13184 -->
     <dependency>
       <groupId>org.opencypher</groupId>
-      <artifactId>okapi-ir</artifactId>
+      <artifactId>spark-cypher</artifactId>
+      <version>${project.parent.version}</version>
+      <scope>test</scope>
+    </dependency>
+
+    <dependency>
+      <groupId>org.opencypher</groupId>
+      <artifactId>spark-cypher</artifactId>
       <version>${project.parent.version}</version>
       <type>test-jar</type>
       <scope>test</scope>
     </dependency>
 
+    <dependency>
+      <groupId>org.opencypher</groupId>
+      <artifactId>okapi-tck</artifactId>
+      <version>${project.parent.version}</version>
+      <type>test-jar</type>
+      <scope>test</scope>
+    </dependency>
+
+    <dependency>
+      <groupId>org.opencypher</groupId>
+      <artifactId>tck</artifactId>
+      <version>${dep.tck.version}</version>
+      <scope>test</scope>
+    </dependency>
   </dependencies>
+
 </project>
